requirements.txt
appdirs==1.4.3
asgiref==3.2.8
beautifulsoup4==4.9.1
boto3==1.12.36
botocore==1.15.36
bs4==0.0.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.7
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-resized==0.3.9
django-rest-framework==0.1.0
django-storages==1.9.1
django-taggit==1.2.0
djangorestframework==3.11.0
docutils==0.15.2
et-xmlfile==1.0.1
filelock==3.0.12
gunicorn==20.0.4
html5==0.0.9
html5lib==1.0.1
idna==2.8
importlib-metadata==1.5.0
ipaddr==2.2.0
jdcal==1.4.1
jmespath==0.9.5
kiwisolver==1.2.0
lockfile==0.12.2
matplotlib==3.2.1
msgpack==0.6.2
numpy==1.18.5
openpyxl==3.0.3
packaging==20.3
pandas==1.0.4
parrot-terminal==0.3
pep517==0.8.2
pep8==1.7.1
Pillow==7.0.0
progress==1.5
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
s3transfer==0.3.3
scipy==1.4.1
seaborn==0.10.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
twilio==6.41.0
urllib3==1.25.8
virtualenv==20.0.13
webencodings==0.5.1
whitenoise==5.0.1
zipp==3.1.0
